Rename ports_config.h to options_ports.h for MPLAB-X projects.
