+++ libUDB/libUDB.c
@@ -32,7 +32,7 @@
 #include "analogs.h"
 #include "events.h"
 #include "osd.h"
-#include "ports_config.h"
+#include "options_ports.h"
 
 #if (USE_I2C1_DRIVER == 1)
 #include "I2C.h"
