+++ libUDB/uart.c
@@ -22,7 +22,7 @@
 #include "libUDB.h"
 #include "oscillator.h"
 #include "uart.h"
-#include "ports_config.h"
+#include "options_ports.h"
 
 #if defined(__dsPIC33E__)
 #include <p33Exxxx.h>
